hw8_model_production/requirements.txt
pandas==1.3.3
scikit-learn==0.23.2
Flask==1.1.2
pytest==6.2.4
pyyaml==5.4.1
click==7.1.2
marshmallow-dataclass==8.5.3
python-dotenv==0.19.1